Updated light.css
cc @simurai 

**borders**
I corrected the colors ```-secondary``` and ```-tertiary``` were pointing to. This is based on how they are used as opposed to how they map to our current css utilities. 

**states**
Replaced ```hl``` with ```state``` in the global variables section. We can keep this group for the app-level ones because some highlights aren't states but for the global ones there are only states so I figured we'd give them a more self-explanatory so they can be reused with confidence. 

**backgrounds**
I removed  ```selected: $blue-500, // `$bg-blue` in``` becuase it's already defined as ```selected-primary-bg: $blue-500,// `$bg-blue` in Primer CSS```